managements moved to manage.sh build system :checkered_flag:
