+++ .vscode/settings.json
@@ -34,8 +34,10 @@
     "شمسی",
     "شنبه",
     "شهریور",
+    "فارسی",
     "فروردین",
     "مرداد",
+    "نسخه",
     "چهارشنبه",
     "یکشنبه"
   ]
