+++ tests/fndict.test.js
@@ -1,0 +1,7 @@
+/* eslint-env jest */
+
+describe('fn dictionary', () => {
+  test('TODO', () => {
+    expect(true).toBe(true)
+  })
+})
